requirements.txt
Flask
flask-cors
geocoder
geopy
joblib
pandas
requests
numpy
scikit-learn==1.3.1
mysql-connector-python